% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBT.perm.R
\name{EBT.perm}
\alias{EBT.perm}
\title{Permutation Test for Energy-based Statistics}
\usage{
EBT.perm(dst, sizes, alpha, B)
}
\arguments{
\item{dst}{A distance matrix of data set}

\item{sizes}{A 2-dimension vector of sample sizes}

\item{alpha}{Significance level}

\item{B}{The number of replicate in permutation test}
}
\value{
A decision on whether the null hypothesis is rejected
}
\description{
Perform Permutation Test for Energy-based Test Statistics
}
\examples{
z<- matrix(rnorm(100),ncol=5)
dst<- L.p.distance(z,2)
EBT.perm(dst,c(10,10),0.05,10)
}
