% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBT.R
\name{EBT}
\alias{EBT}
\title{Energy-based Test}
\usage{
EBT(dst, ix, sizes)
}
\arguments{
\item{dst}{A distance matrix of data set}

\item{ix}{A permutation of row indices of dst}

\item{sizes}{A 2-dimension vector of sample sizes}
}
\value{
A 1-dimension numeric value of test statistics
}
\description{
Evaluate Energy-based Test Statistics
}
\examples{
z<- matrix(rnorm(100),ncol=5)
dst<- L.p.distance(z,2)
EBT(dst,1:20,c(10,10))
}
