% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GST.perm.R
\name{GST.perm}
\alias{GST.perm}
\title{Permutation Test for Graph-based Test Statistics}
\usage{
GST.perm(dst, sizes, Q, alpha, B)
}
\arguments{
\item{dst}{A distance matrix of data set}

\item{sizes}{A 2-dimension vector of sample sizes}

\item{Q}{The threshold in graph construction}

\item{alpha}{Significance level}

\item{B}{The number of replicate in permutation test}
}
\value{
A decision on whether the null hypothesis is rejected
}
\description{
Perform Permutation Test for Graph-based Test Statistics
}
\examples{
z<- matrix(rnorm(100),ncol=5)
dst<- L.p.distance(z,2)
GST.perm(dst,c(10,10),2,0.05,10)
}
