% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTT.perm.R
\name{HTT.perm}
\alias{HTT.perm}
\title{Permutation Test for Hotelling's T-squared Test Statistics}
\usage{
HTT.perm(z, sizes, alpha, B)
}
\arguments{
\item{z}{A matrix of data set (each row is an observation)}

\item{sizes}{A 2-dimension vector of sample sizes}

\item{alpha}{Significance level}

\item{B}{The number of replicate in permutation test}
}
\value{
A decision on whether the null hypothesis is rejected
}
\description{
Perform Permutation Test for Hotelling's T-squared Test Statistics
}
\examples{
z<- matrix(rnorm(100),ncol=5)
HTT.perm(z,c(10,10),0.05,10)
}
