% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTT.R
\name{HTT}
\alias{HTT}
\title{Hotelling's T-squared Test}
\usage{
HTT(z, ix, sizes)
}
\arguments{
\item{z}{A matrix of data set (each row is an observation)}

\item{ix}{A permutation of row indices of z}

\item{sizes}{A 2-dimension vector of sample sizes}
}
\value{
A 1-dimension numeric value of test statistics
}
\description{
Evaluate Hotelling's T-squared Test Statistics
}
\examples{
z<- matrix(rnorm(100),ncol=5)
HTT(z,1:20,c(10,10))
}
