% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L.p.distance.R
\name{L.p.distance}
\alias{L.p.distance}
\title{Lp Distance}
\usage{
L.p.distance(x, p)
}
\arguments{
\item{x}{A matrix of data set (each row is an observation)}

\item{p}{The specific form of distance}
}
\value{
A distance matrix of the chosen distance form
}
\description{
Calculating the designated form of distance matrix
}
\examples{
z<- matrix(rnorm(100),ncol=5)
L.p.distance(z,2)
}
