% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNT.R
\name{NNT}
\alias{NNT}
\title{Nearest Neighbot Test}
\usage{
NNT(z, ix, sizes, R)
}
\arguments{
\item{z}{A matrix of data set (each row is an observation)}

\item{ix}{A permutation of row indices of z}

\item{sizes}{A 2-dimension vector of sample sizes}

\item{R}{The number of neighborhood to be evaluated}
}
\value{
A 1-dimension numeric value of test statistics
}
\description{
Evaluate Nearest Neighbor Test Statistics
}
\examples{
z<- matrix(rnorm(100),ncol=5)
NNT(z,1:20,c(10,10),3)
}
